Move guild scheduled event stuff to a guild scheduled event directory
