+++ src/events/guild_scheduled_event/guildScheduledEventUserAdd.ts
