+++ src/events/guild_scheduled_event/guildScheduledEventUserRemove.ts
