+++ src/events/index.ts
@@ -18,10 +18,10 @@ export { guildMemberUpdate } from './guild_member/guildMemberUpdate.js';
 
 export { guildMemberVoiceUpdate } from './guild_member/guildMemberVoiceUpdate.js';
 
-export { guildScheduledEventUpdate } from './attendance/guildScheduledEventUpdate.js';
+export { guildScheduledEventUpdate } from './guild_scheduled_event/guildScheduledEventUpdate.js';
 
-export { guildScheduledEventUserAdd } from './attendance/guildScheduledEventUserAdd.js';
+export { guildScheduledEventUserAdd } from './guild_scheduled_event/guildScheduledEventUserAdd.js';
 
-export { guildScheduledEventUserRemove } from './attendance/guildScheduledEventUserRemove.js';
+export { guildScheduledEventUserRemove } from './guild_scheduled_event/guildScheduledEventUserRemove.js';
 
 export { voiceStateUpdate } from './attendance/voiceStateUpdate.js';
